+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Use new dependency resolver for template contract - [#325](https://github.com/paritytech/cargo-contract/pull/325)
+- Do not strip out panic messages in debug builds - [#326](https://github.com/paritytech/cargo-contract/pull/326)
 
 ## [0.13.1] - 2021-08-03