+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.14.0] - 2021-08-12
+
 ### Added
 -  Add option for JSON formatted output - [#324](https://github.com/paritytech/cargo-contract/pull/324)